+++ i18n/utils/gettext/__init__.py
@@ -1,16 +1,16 @@
 from .shortcuts import (
     set_language,
     set_domain,
-    get_language
+    get_language,
+    translate,
 )
 
-from .contexts import (
-    language
-)
+from .contexts import language
 
 __all__ = [
-    'set_domain',
-    'set_language',
-    'get_language',
-    'language'
+    "set_domain",
+    "set_language",
+    "get_language",
+    "language",
+    "translate",
 ]