+++ i18n/utils/imputils/__init__.py
@@ -1,5 +1,3 @@
 from .utils import import_module
 
-__all__ = [
-    'import_module'
-]
+__all__ = ["import_module"]