+++ i18n/utils/imputils/utils.py
@@ -8,19 +8,19 @@ from types import ModuleType
 @cache
 def import_module(path: str) -> ModuleType:
     if os.path.isdir(path):
-        path = os.path.join(path, '__main__.py')
+        path = os.path.join(path, "__main__.py")
 
     directory = os.path.dirname(path)
     sys.path.append(directory)
     try:
         if not os.path.exists(path):
-            raise ImportError(f'Module not found at: {path}')
+            raise ImportError(f"Module not found at: {path}")
         *_, filename = os.path.split(path)
-        spec = importlib.util.spec_from_file_location(filename.replace('.py', ''), path)
+        spec = importlib.util.spec_from_file_location(filename.replace(".py", ""), path)
         imported_module = importlib.util.module_from_spec(spec)
         spec.loader.exec_module(imported_module)
         return imported_module
     except Exception:
-        raise ImportError(f'Failed to import module: {path}')
+        raise ImportError(f"Failed to import module: {path}")
     finally:
         sys.path.remove(directory)