+++ i18n/utils/potfile/__init__.py
@@ -1,7 +1,4 @@
 from .potfile import POTFile
 from .constants import TEMPLATES
 
-__all__ = [
-    'POTFile',
-    'TEMPLATES'
-]
+__all__ = ["POTFile", "TEMPLATES"]