+++ i18n/utils/potfile/types.py
@@ -1,3 +1,3 @@
 from typing import Literal, NewType
 
-TextFileMode = NewType('TextFileMode', Literal['r', 'w', 'a'])
+TextFileMode = NewType("TextFileMode", Literal["r", "w", "a"])