+++ tests/i18n/test_translatable_enum.py
@@ -4,23 +4,31 @@ from i18n import TranslatableEnum, set_domain, set_language
 
 
 def test_translatable_enums():
-    set_language('uk')
-    set_domain('app', './tests/languages')
+    set_language("uk")
+    set_domain("app", "./tests/languages")
 
     class MyEnum(TranslatableEnum):
-        HELLO = 'hello'
-        WORLD = 'world'
+        HELLO = "hello"
+        WORLD = "world"
 
-    assert MyEnum.WORLD.value == 'Світ'
+    assert MyEnum.WORLD.value == "Світ"
 
 
 def test_translate_to_language():
-    set_language('en')
-    set_domain('app', './tests/languages')
+    set_language("en")
+    set_domain("app", "./tests/languages")
 
     class MyEnum(TranslatableEnum):
-        HELLO = 'hello'
-        WORLD = 'world'
+        HELLO = "hello"
+        WORLD = "world"
 
-    assert MyEnum.WORLD.language('uk') == 'Світ'
-    assert MyEnum.HELLO.value == 'hello'
+    assert MyEnum.WORLD.language("uk") == "Світ"
+    assert MyEnum.HELLO.value == "hello"
+
+
+def test_string_comparing():
+    class MyEnum(TranslatableEnum):
+        HELLO = "hello"
+        WORLD = "world"
+
+    assert MyEnum.HELLO == str(MyEnum.HELLO)